refactor: GetCrateUsage was incorrectly rounding storage sizes, leading to incorrect data visualization on the client.

+++ CloudCrate.Application/DTOs/Crate/CrateUsageDto.cs
@@ -4,9 +4,9 @@ namespace CloudCrate.Application.DTOs.Crate;
 
 public class CrateUsageDto
 {
-    public int TotalUsedStorage { get; set; }
-    public int StorageLimit { get; set; }
-    public int RemainingStorage => StorageLimit - TotalUsedStorage;
+    public double TotalUsedStorage { get; set; }
+    public double StorageLimit { get; set; }
+    public double RemainingStorage => StorageLimit - TotalUsedStorage;
 
     public List<FileTypeBreakdownDto> BreakdownByType { get; set; } = new();
 }
